x86-64 rewrite

+++ include/kernel/types.h
@@ -3,8 +3,8 @@
 
 #include <inttypes.h>
 
-typedef uint32_t phys_addr_t;
-typedef uint32_t virt_addr_t;
+typedef uint64_t phys_addr_t;
+typedef uint64_t virt_addr_t;
 
 #define __pack __attribute__((packed))
 #define __packalign(n) __attribute__((packed, aligned(n)))
